feat(sbb-seat-reservation): fix build errors - i18n translations moved to common without subdir - quick (temporary) fix for @event type definitions
