+++ src/elements-experimental/seat-reservation/common/i18n.ts
